dagster_workflow/requirements.txt
requests==2.30.0
patool== 1.12
beautifulsoup4==4.11.2
pandas==1.1.5
selenium==4.9
webdriver-manager==3.8.6
undetected-chromedriver==2.1.1
dagit
dagster==1.5.3
